Add files via upload
This is the Day of the Week application. By running this application, you can select any date from the calendar. If it is a weekday, it will display a stressful worker because you will have to work. However, if it is a weekend, it will display a vacation on the beach because you do not have work so you can relax.