Add files via upload
This is an application I created called Hospital. The main program has a menu that allows the user to enter a type of surgery, a type of medication, number of days of hospital stay, and check the patient out of the hospital. When the patient checks out, the total charges should be displayed.