Add files via upload
This is an application I created called Grade. In this application, you can calculate the grade of a student by adding five different grades together and the application will average them and display the average of the student. Depending on the grade of the average it will display the number, and letter of the grade. Also, it will display a face depending on the grade of the class.